Move obo importing stuff into its own folder
